# swagger-spec.json
{"openapi":"3.0.0","paths":{"/users":{"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":"An array of all users","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PublicUser"}}}}}},"tags":["User"]}},"/users/id/{id}":{"get":{"operationId":"UserController_findById","parameters":[],"responses":{"200":{"description":"One user with the specified id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrivateUser"}}}}},"tags":["User"]},"patch":{"operationId":"UserController_updateOne","summary":"","description":"Update a user","parameters":[],"requestBody":{"required":true,"description":"Partial User object with fields to be updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/updateUserDto"}}}},"responses":{"200":{"description":"The new User document","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrivateUser"}}}}},"tags":["User"]},"delete":{"operationId":"UserController_deleteOne","summary":"","description":"Delete a user","parameters":[],"responses":{"200":{"description":"The deleted User document","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrivateUser"}}}}},"tags":["User"]}},"/users/email/{email}":{"get":{"operationId":"UserController_findByEmail","parameters":[{"name":"email","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"One user with the specified email","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["User"]}},"/rooms":{"get":{"operationId":"RoomController_findAll","summary":"","description":"Get all existing rooms","parameters":[],"responses":{"default":{"description":"List of all rooms","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UnpopulatedRoom"}}}}}},"tags":["Room"]},"post":{"operationId":"RoomController_addOne","summary":"","description":"Creates a new room","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoomDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PopulatedRoom"}}}}},"tags":["Room"]}},"/rooms/id/{id}":{"get":{"operationId":"RoomController_findById","summary":"","description":"Get one room with specified id","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PopulatedRoom"}}}}},"tags":["Room"]}},"/rooms/name/{name}":{"get":{"operationId":"RoomController_findByName","summary":"","description":"Get one room with specified name","parameters":[{"name":"name","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PopulatedRoom"}}}}},"tags":["Room"]}},"/rooms/user/{user}":{"get":{"operationId":"RoomController_findByUser","summary":"","description":"Get all rooms for a specific User","parameters":[],"responses":{"default":{"description":"A list of all rooms where Room.users includes User","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PopulatedRoom"}}}}}},"tags":["Room"]}},"/rooms/owned/{user}":{"get":{"operationId":"RoomController_findByOwner","summary":"","description":"Get all rooms owned by a specific User","parameters":[],"responses":{"default":{"description":"A list of all rooms where Room.owner === User","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PopulatedRoom"}}}}}},"tags":["Room"]}},"/rooms/id/{id}/join":{"post":{"operationId":"RoomController_addUser","summary":"","description":"Adds a User to this room","parameters":[],"responses":{"default":{"description":"The new Room document(unpopulated)","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UnpopulatedRoom"}}}}}},"tags":["Room"]}},"/rooms/id/{id}/leave":{"post":{"operationId":"RoomController_removeUser","summary":"","description":"Removes a User from this room","parameters":[],"responses":{"default":{"description":"The new Room document(unpopulated)","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UnpopulatedRoom"}}}}}},"tags":["Room"]}},"/rooms/id/{id}/messages":{"post":{"operationId":"RoomController_addMessage","summary":"","description":"Adds a message to this room","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendMessageDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnpopulatedMessage"}}}}},"tags":["Room"]},"patch":{"operationId":"RoomController_editMessage","summary":"","description":"Edit text of the message","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMessageDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnpopulatedMessage"}}}}},"tags":["Room"]},"get":{"operationId":"RoomController_getMessages","summary":"","description":"Get all messages of this room","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"from_date","required":false,"in":"path","description":"Date string, define to select only messages created after this date","schema":{"type":"string"}},{"name":"to_date","required":false,"in":"path","description":"Date string, define to select only messages created before this date","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PopulatedMessage"}}}}}},"tags":["Room"]}},"/rooms/id/{id}/messages/search":{"get":{"operationId":"RoomController_searchMessages","summary":"","description":"Search messages in this room","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"author","required":false,"in":"query","description":"ObjectId of a user that needs to be the author of resulting messages","schema":{"type":"string"}},{"name":"text","required":false,"in":"query","description":"A string that should be contained inside all resulting messages","schema":{"type":"string"}},{"name":"from_date","required":false,"in":"query","description":"ISO Date string representing the start of the search period","schema":{"type":"string"}},{"name":"to_date","required":false,"in":"query","description":"ISO Date string representing the end of the search period","schema":{"type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PopulatedMessage"}}}}}},"tags":["Room"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/register":{"post":{"operationId":"AuthController_register","summary":"","description":"Create a new user","parameters":[],"requestBody":{"required":true,"description":"The user to be created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/registerUserDto"}}}},"responses":{"201":{"description":"The new User document","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrivateUser"}}}}},"tags":["auth"]}},"/auth/self":{"post":{"operationId":"AuthController_private","summary":"","description":"Returns current authenticated user","parameters":[],"responses":{"201":{"description":""}},"tags":["auth"]}},"/":{"get":{"operationId":"AppController_getServerInfo","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/"}}}}}}}},"info":{"title":"Simple chat","description":"A basic chat","version":"0.0.1","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"Role":{"type":"string","enum":["user","admin"]},"PublicUser":{"type":"object","properties":{"_id":{"type":"string"},"name":{"type":"string"},"role":{"description":"User role, will be used for protected routes","$ref":"#/components/schemas/Role"}},"required":["_id","name","role"]},"PrivateUser":{"type":"object","properties":{"_id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string","description":"User email, private"},"password":{"type":"string","description":"Password hash, private"},"role":{"description":"User role, will be used for protected routes","$ref":"#/components/schemas/Role"}},"required":["_id","name","email","password","role"]},"User":{"type":"object","properties":{"_id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"}},"required":["_id","name","role"]},"updateUserDto":{"type":"object","properties":{"name":{"type":"string","description":"New username, from 1 to 25 characters"},"email":{"type":"string"},"role":{"description":"User role, will be used for protected routes","$ref":"#/components/schemas/Role"}},"required":["name","email","role"]},"UnpopulatedRoom":{"type":"object","properties":{"_id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"users":{"description":"Array of ObjectId-s of users that are members in this room","type":"array","items":{"type":"string"}},"owner":{"type":"string","description":"ObjectId of a User that owns this room"}},"required":["_id","name","description","users","owner"]},"CreateRoomDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"owner":{"type":"string","description":"ObjectId of the User that owns this room"}},"required":["name","owner"]},"PopulatedRoom":{"type":"object","properties":{"_id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"users":{"description":"Array of User objects that are members in this room","type":"array","items":{"$ref":"#/components/schemas/PublicUser"}},"owner":{"description":"The User that owns this room","allOf":[{"$ref":"#/components/schemas/PublicUser"}]}},"required":["_id","name","description","users","owner"]},"SendMessageDto":{"type":"object","properties":{"text":{"type":"string"},"forwardOf":{"type":"string","description":"An ObjectId of another message, if defined, thus message is a \"reply\" to that message"},"author":{"type":"string","description":"An ObjectId of the User that sent this message"}},"required":["text","author"]},"UnpopulatedMessage":{"type":"object","properties":{"_id":{"type":"string"},"text":{"type":"string"},"author":{"type":"string"},"room":{"type":"string"},"ForwardOf":{"type":"string","description":"If defined, this message is a reply to the message with specified ObjectId"}},"required":["_id","text","author","room"]},"ObjectID":{"type":"object","properties":{}},"UpdateMessageDto":{"type":"object","properties":{"id":{"description":"ObjectId string of the message to be updated","allOf":[{"$ref":"#/components/schemas/ObjectID"}]},"text":{"type":"string"}},"required":["id","text"]},"PopulatedMessage":{"type":"object","properties":{"_id":{"type":"string"},"text":{"type":"string"},"author":{"$ref":"#/components/schemas/PublicUser"},"room":{"type":"string"},"ForwardOf":{"description":"If defined, this message is a reply to the specified message","allOf":[{"$ref":"#/components/schemas/UnpopulatedMessage"}]}},"required":["_id","text","author","room"]},"registerUserDto":{"type":"object","properties":{"name":{"type":"string","description":"Username, from 1 to 25 characters"},"email":{"type":"string","description":"A valid email. Should be unique for every User"},"password":{"type":"string","description":"A password, miminum 5 chars, will be hashed"},"role":{"description":"User role, will be used for protected routes","$ref":"#/components/schemas/Role"}},"required":["name","email","password","role"]}}}}